🐛: devcontainer does not take the project name from the existing .env file.
When the .devcontainer starts automatically, it takes "--project-name as 'project folder name' and its definition in the .env file in the current folder is not applied."
This problem also causes similar projects with the same name to run in the same container.

Additionally, .devcontaine does not reference $ in the .env in the home directory during initialization.